ci: enable ci fail for xlspike run and manual job
see https://docs.gitlab.com/ci/jobs/job_control/#types-of-manual-jobs

In optional manual jobs:

allow_failure is true, which is the default setting for jobs that have when: manual defined outside of rules.
The status does not contribute to the overall pipeline status. A pipeline can succeed even if all of its manual jobs fail.
In blocking manual jobs:

allow_failure is false, which is the default setting for jobs that have when: manual defined inside rules.
The pipeline stops at the stage where the job is defined. To let the pipeline continue running, run the manual job.
Merge requests in projects with Pipelines must succeed enabled can’t be merged with a blocked pipeline.
The pipeline shows a status of blocked.

Signed-off-by: Huaqi Fang <578567190@qq.com>